Maximum Likelihood Estimation of a TVGEV Model

Maximum Likelihood Estimation of a TVGEV model.

MLE.TVGEV(object, y = NULL,
          psi0 = NULL, estim = c("optim", "nloptr"),
          coefLower = if (estim != "nloptr") numeric(0) else c("xi_0" = -0.9),
          coefUpper = if (estim != "nloptr") numeric(0) else c("xi_0" = 2.0),
          parTrack = FALSE)

Arguments

object

A (possibly incomplete) TVGEV object.

y

A numeric vector giving the response to be used.

psi0

Numeric vector of initial values for the parameters.

estim

Character giving the optimisation to be used.

coefLower, coefUpper

Numeric vectors or NULL giving bounds on the parameters. These are used overwrite the "normal" values which are -Inf and Inf. These arguments are only used when estim is "nloptr".

parTrack

Logical. If TRUE, all parameters at which an evaluation of the log-Likelihood will be stored and returned.

Value

A list with elements that can be copied into those of a TVGEV object.

Note

The box constraints on parameters defined by coefLower and coefUpper should not be active at the optimum. If this happens, the inference will be misleading. Note that these constaints can be used to set some parameters to a given value by using the same value for the lower and the upper bound. In this case the fact that the inference is misleading is clear from the fact that the number of parameters in the object is wrong.

Caution

For now it is assumed that the shape parameter is constant, hence is equal to the element named xi_0 in the parameter vector. This assumption is used to set the bounds on the shape parameter when estim is set to "nloptr": the lower bound is -0.9 and the upper bound is 2.0.

Author

Yves Deville