Non-Stationary GEV Time Series

Utility functions for Non-Stationary time series with GEV margins.

Details

The DESCRIPTION file: This package was not yet installed at build time.

This package is dedicated to some Non-Stationary Extreme Value models with Generalised Extreme Value (GEV) marginal distributions. Such models are often used to investigate climate change from series of annual maxima.

An emphasis is put on Profile Likelihood inference on the parameters, quantiles or return levels. For that aim constrained optimisations are used, relying on the nloptr package.

The package provides most of the classical S3 methods useful to cope with fitted model objects, such as predict, coef, vcov, confint, residuals, logLik, ...

"NSGEV" This class was intended to be the main class when the package was created. As a major difference with other EV packages, the model parameters can be used in several of the GEV parameters loc, scale and shape. A typical example is when a scale proportional to the location needs to be specified.

"TVGEV" This class is devoted to the special case of Time Varying (TV) models using only functions of the time as covariates. Specialised methods make sense to plot, predict and more.

Most efforts have been put on the "TVGEV" class. In the future, the "NSGEV" class could be discarded.

Author

Yves Deville [cre, aut] (ORCID: <https://orcid.org/0000-0002-1233-488X>)

Maintainer: Yves Deville <deville.yves@alpestat.com>

References

Coles, S. (2001) Introdution to Statistical Modelling of Extreme Values, Springer.

Examples

library(NSGEV)
methods(class = "TVGEV")
#>  [1] anova         autoplot      bs            cdf           cdfMaxFun    
#>  [6] coef          confint       density       influence     logLik       
#> [11] mean          modelMatrices moment        plot          predict      
#> [16] print         profLik       psi2theta     quantMax      quantMaxFun  
#> [21] quantile      residuals     simulate      summary       vcov         
#> see '?methods' for accessing help and source code