Value of the Probability Density of the Maximum over a Period of Time

Compute the probability density of the maximum over a period of time.

dMax.TVGEV(
  object,
  x,
  date = NULL,
  psi = NULL,
  deriv = FALSE,
  derx = FALSE,
  ...
)

Arguments

object

A TVGEV object.

x

A vector of values at which the density is to be evaluated.

date

A vector that can be coerced to the class "Date" which defines the period of time.

psi

Optional vector of parameters. By default, the estimated vector of parameters coef(object) is used.

deriv

Logical. If TRUE the gradient of the density will be computed and returned as the "gradient" attribute of the result.

derx

Logical, If TRUE the derivative of the density w.r.t. x will be computed and returned as the "derx" attribute of the result.

...

Not used yet.

Value

A vector of values for the probability density. Beside the gradient (w.r.t. the parameters) returned as the attribute named "gradient", the derivative w.r.t. x is also returned as the attribute "derx" of the result.

Caution

This function may be renamed in a future version.

References

See the Computing Details vignette of the package for the computation of the derivatives.