Find the complete vector of NSGEV parameters from a partial vector and a Return Level

Find the complete vector of NSGEV parameters from a partial vector and a Return Level.

rho2psi(rho, nm1, psi_m1, model, data = NULL, type = "expect")

Arguments

rho

A fixed Return Level (RL).

nm1

Name of the parameter which is adjusted.

psi_m1

Vector of NSGEV parameters with its element nm1 removed.

model

The NSGEV model.

data

Data frame of covariates.

type

The type of Return Level.

Value

A vector \(\boldsymbol{\psi}\) of NSGEV parameters.

Details

In a Non-Sationary framework, a Return Level (RL) rho relates to a given value of the covariates or to a distribution of the covariates. This distribution is assumed here to be given by the observations in data, assuming that the corresponding Return Period is equal to the number of observations understood as a multiple of the block duration. When type = "expect" the RL is the value \(\rho\) for which the random number of exceedances over \(\rho\) has unit expectation.

This function is a technical function to profile the likelihood. One of the NSGEV model parameter, say \(\psi_1\), is adjusted to reach the given value RL, the other elements of \(\psi\) being given and fixed. These later elements form a vector \(\boldsymbol{\psi}_{-1}\) with length \(p-1\), where \(p\) is the number of model parameters. Now for a given value of \(rho\), the value of the profile log-likelihood \(\ell(\rho)\) is obtained by maximising the log-likelihood w.r.t \(\boldsymbol{\psi}_{-1}\).

Author

Yves Deville

Examples

df <- data.frame(t = 1:10)
fit <- NSGEV(formulas = list("loc" = ~ alpha + beta * t,  "scale" = ~ delta, "shape" = ~ xi),
                             data = df)
df.new <- data.frame(t = 11:20)
psi <- c("alpha" = 1, "beta" = 0.01, "delta" = 0.6, "xi" = 0.06)
rho2psi(rho = 30, nm1 = "alpha", psi_m1 = psi[-1], model = fit, data = df.new)
#>    alpha     beta    delta       xi 
#> 28.39881  0.01000  0.60000  0.06000 
rho2psi(rho = 40, nm1 = "alpha", psi_m1 = psi[-1], model = fit, data = df.new)
#>    alpha     beta    delta       xi 
#> 38.39881  0.01000  0.60000  0.06000