Test that a Provided Graident, Jacobian or Hessian is in Accordance with a Numeric Hessian

Test that a provided object containing derivatives is in good accordance with another similar object containing numeric derivatives.

testNumDeriv(der, derNum, PREC = 0.05, type = c("gradient", "hessian"))

Arguments

der

The provided vector, matrix or array of derivatives to be checked.

derNum

The numeric vector, matrix or array of derivatives.

PREC

A precision parameter. A value of 0.05 seems good in practice.

Value

A logical which is TRUE if the two objects are close enough.

Details

One could think of requiring a small absolute difference or a small relative difference between the results. However, in practice it seems that the absolute difference can be large even with correct derivatives because these contain large values. Also the relative difference can be large if the derivatives contain values that are small in absolute value. However, if one of these two condition is fulfilled, the result must be correct because none of the condition can be true by chance.